Task #208808 refactor: Changes for Cohort middleware
