+++ src/src/beneficiaries/beneficiaries.module.ts
@@ -1,18 +1,17 @@
-import { MiddlewareConsumer, Module, NestModule } from '@nestjs/common';
+import { Module } from '@nestjs/common';
 import { UserModule } from 'src/user/user.module';
 import { BeneficiariesController } from './beneficiaries.controller';
 
 import { HttpModule } from '@nestjs/axios';
 import { S3Module } from 'src/services/s3/s3.module';
-import { AuthMiddleware } from '../common/middlewares/authmiddleware';
+import { UploadFileModule } from 'src/upload-file/upload-file.module';
+import { EnumModule } from '../enum/enum.module';
 import { HasuraModule } from '../hasura/hasura.module';
 import { HelperModule } from '../helper/helper.module';
-import { EnumModule } from '../enum/enum.module';
 import { HasuraModule as HasuraModuleFromServices } from '../services/hasura/hasura.module';
 import { KeycloakModule } from '../services/keycloak/keycloak.module';
-import { BeneficiariesService } from './beneficiaries.service';
-import { UploadFileModule } from 'src/upload-file/upload-file.module';
 import { BeneficiariesCoreService } from './beneficiaries.core.service';
+import { BeneficiariesService } from './beneficiaries.service';
 
 @Module({
 	imports: [
@@ -31,12 +30,11 @@ import { BeneficiariesCoreService } from './beneficiaries.core.service';
 	providers: [BeneficiariesService, BeneficiariesCoreService],
 	exports: [BeneficiariesService, BeneficiariesCoreService],
 })
-// export class BeneficiariesModule implements NestModule {
-// 	configure(consumer: MiddlewareConsumer) {
-// 		consumer.apply(AuthMiddleware).forRoutes('*');
-		
-// 	
-//}
-//}
-export class BeneficiariesModule{}
 
+/*export class BeneficiariesModule implements NestModule {
+	configure(consumer: MiddlewareConsumer) {
+		consumer.apply(AuthMiddleware).forRoutes('*');
+    }
+}*/
+
+export class BeneficiariesModule {}
