+++ src/src/comments/comments.module.ts
@@ -1,6 +1,5 @@
-import { MiddlewareConsumer, Module, NestModule } from '@nestjs/common';
+import { Module } from '@nestjs/common';
 import { UserModule } from 'src/user/user.module';
-import { AuthMiddleware } from '../common/middlewares/authmiddleware';
 import { HasuraModule } from '../hasura/hasura.module';
 import { HasuraModule as HasuraModuleFromServices } from '../services/hasura/hasura.module';
 import { CommentsController } from './comments.controller';
@@ -11,9 +10,11 @@ import { CommentsService } from './comments.service';
   controllers: [CommentsController],
   providers: [CommentsService]
 })
-// export class CommentsModule implements NestModule {
-//   configure(consumer: MiddlewareConsumer) {
-// 		consumer.apply(AuthMiddleware).forRoutes(CommentsController);
-// 	}
-// }
+
+/*export class CommentsModule implements NestModule {
+  configure(consumer: MiddlewareConsumer) {
+		consumer.apply(AuthMiddleware).forRoutes(CommentsController);
+	}
+}*/
+
 export class CommentsModule{}
