+++ src/src/facilitator/facilitator.module.ts
@@ -1,18 +1,16 @@
-import { MiddlewareConsumer, Module, NestModule } from '@nestjs/common';
 import { HttpModule } from '@nestjs/axios';
+import { Module } from '@nestjs/common';
 
-import { HasuraModule } from '../services/hasura/hasura.module';
-import { EnumModule } from '../enum/enum.module';
 import { UserModule } from 'src/user/user.module';
+import { EnumModule } from '../enum/enum.module';
 import { AuthModule } from '../modules/auth/auth.module';
+import { HasuraModule } from '../services/hasura/hasura.module';
 import { S3Module } from '../services/s3/s3.module';
 
-import { FacilitatorService } from './facilitator.service';
-import { FacilitatorController } from './facilitator.controller';
-import { AuthMiddleware } from '../common/middlewares/authmiddleware';
 import { UploadFileModule } from 'src/upload-file/upload-file.module';
+import { FacilitatorController } from './facilitator.controller';
 import { FacilitatorCoreService } from './facilitator.core.service';
-import { CohortMiddleware } from 'src/common/middlewares/cohort_middleware/cohort_middleware.middleware'; 
+import { FacilitatorService } from './facilitator.service';
 
 @Module({
 	imports: [
@@ -24,14 +22,16 @@ import { CohortMiddleware } from 'src/common/middlewares/cohort_middleware/cohor
 		S3Module,
 		UploadFileModule,
 	],
-	providers: [FacilitatorService,FacilitatorCoreService],
+	providers: [FacilitatorService, FacilitatorCoreService],
 	controllers: [FacilitatorController],
 	//exports: [FacilitatorCoreService,FacilitatorService],
 })
-// export class FacilitatorModule implements NestModule {
-// 	configure(consumer: MiddlewareConsumer) {
-// 		consumer.apply(AuthMiddleware).forRoutes('*');
-// 		consumer.apply(CohortMiddleware).forRoutes('*');
-// 	}
-// }
-export class FacilitatorModule{}
+
+/*export class FacilitatorModule implements NestModule {
+	configure(consumer: MiddlewareConsumer) {
+		consumer.apply(AuthMiddleware).forRoutes('*');
+		consumer.apply(CohortMiddleware).forRoutes('*');
+	}
+}*/
+
+export class FacilitatorModule {}
