+++ src/src/modules/auth/auth.module.ts
@@ -6,7 +6,7 @@ import { KeycloakModule } from 'src/services/keycloak/keycloak.module';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 import { UserModule } from 'src/user/user.module';
-import { AuthMiddleware } from 'src/common/middlewares/authmiddleware';
+import { AuthMiddleware } from 'src/common/middlewares/auth.middleware';
 
 @Module({
 	imports: [
